refactor(test setup): moved mongomemoryserver into test root for reusability
